deleted old viewcontroller, rename new in old
