+++ XO-game/GameViewController.swift
@@ -1,31 +1,90 @@
 //
-//  GameViewController.swift
+//  StartViewController.swift
 //  XO-game
 //
-//  Created by Evgeny Kireev on 25/02/2019.
-//  Copyright © 2019 plasmon. All rights reserved.
+//  Created by Nikkru on 23.03.2022.
+//  Copyright © 2022 plasmon. All rights reserved.
 //
 
 import UIKit
 
 class GameViewController: UIViewController {
-
-    @IBOutlet var gameboardView: GameboardView!
-    @IBOutlet var firstPlayerTurnLabel: UILabel!
-    @IBOutlet var secondPlayerTurnLabel: UILabel!
-    @IBOutlet var winnerLabel: UILabel!
-    @IBOutlet var restartButton: UIButton!
     
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        gameboardView.onSelectPosition = { [weak self] position in
-            guard let self = self else { return }
-            self.gameboardView.placeMarkView(XView(), at: position)
+    var mainView: GameMainView { return self.view as! GameMainView }
+    
+    private let gameboard = Gameboard()
+    private lazy var referee = Referee(gameboard: gameboard)
+    private var currentState: GameStateProtocol! {
+        didSet { self.currentState.begin()
         }
     }
     
-    @IBAction func restartButtonTapped(_ sender: UIButton) {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        self.goToFirstState()
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+    }
+    
+    override func loadView() {
         
+        self.view = GameMainView(frame: UIScreen.main.bounds)
+        
+        mainView.gameboardView.onSelectPosition = { [weak self] position in
+            guard let self = self else { return }
+            
+            //            self.mainView.gameboardView.placeMarkView(XView(), at: position)
+            self.currentState.addMark(at: position)
+            if self.currentState.isCompleted {
+                self.goToNextState() }
+        }
+        
+        mainView.onAddGameAction = { [weak self] in
+            self?.addRestartButtonAction()
+        }
+    }
+    
+    private func goToFirstState() {
+        
+        self.currentState = PlayerInputState(
+            player: .first,
+            gameViewController: self,
+            gameboard: gameboard,
+            gameboardView: mainView.gameboardView
+        )
+    }
+    private func goToNextState() {
+        
+        if let winner = self.referee.determineWinner() {
+            self.currentState = GameEndedState(winner: winner, gameViewController: self)
+            return
+        }
+        
+        //  проверка условия для ничейного результата
+        if GameboardSize.fieldsCount == MarkView.stepsCount {
+            let gameEnded = GameEndedState(winner: nil, gameViewController: self)
+            gameEnded.begin()
+
+            
+            return
+        }
+        
+        if let playerInputState = currentState as? PlayerInputState {
+            self.currentState = PlayerInputState(
+                player: playerInputState.player.next,
+                gameViewController: self,
+                gameboard: gameboard,
+                gameboardView: mainView.gameboardView
+            )
+        }
+    }
+    
+    func addRestartButtonAction() {
+        
+        mainView.gameboardView.clear()
+        gameboard.clear()
+        goToFirstState()
     }
 }
-
